requirements.txt
discord.py==2.5.0
python-dotenv==1.0.1
palworld-api==1.0.3
gamercon-async==1.0.7
aiohttp==3.10.10
aiosqlite==0.20.0
aiocache==0.12.3